Removing the pox seurity service component and fixed the requested changes
